Move AtomCertVerifier to atom/browser/net
