+++ filenames.gypi
@@ -131,8 +131,6 @@
       'atom/browser/atom_download_manager_delegate.h',
       'atom/browser/atom_browser_main_parts.cc',
       'atom/browser/atom_browser_main_parts.h',
-      'atom/browser/atom_cert_verifier.cc',
-      'atom/browser/atom_cert_verifier.h',
       'atom/browser/atom_browser_main_parts_mac.mm',
       'atom/browser/atom_browser_main_parts_posix.cc',
       'atom/browser/atom_javascript_dialog_manager.cc',
@@ -175,6 +173,8 @@
       'atom/browser/net/asar/asar_protocol_handler.h',
       'atom/browser/net/asar/url_request_asar_job.cc',
       'atom/browser/net/asar/url_request_asar_job.h',
+      'atom/browser/net/atom_cert_verifier.cc',
+      'atom/browser/net/atom_cert_verifier.h',
       'atom/browser/net/atom_url_request_job_factory.cc',
       'atom/browser/net/atom_url_request_job_factory.h',
       'atom/browser/net/http_protocol_handler.cc',
